requirements.txt
azure-common==1.1.27
azure-core==1.18.0
azure-functions==1.7.2
azure-servicebus==7.3.3
azure-storage-blob==12.9.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
cryptography==3.4.8
idna==3.2
isodate==0.6.0
msrest==0.6.21
oauthlib==3.1.1
pycparser==2.20
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
uamqp==1.4.2
urllib3==1.26.6
numpy==1.21.2
pandas==1.3.3
python-dateutil==2.8.2
pytz==2021.1
tabulate==0.8.9

